agri2018 move code to subfolders energy_demand, energy_source
