+++ generatorcore/agri2018/energy_demand/energy.py
@@ -1,0 +1,9 @@
+# pyright: strict
+
+from dataclasses import dataclass
+
+
+@dataclass(kw_only=True)
+class Energy:
+    # Used by p_operation, p_operation_elec_heatpump
+    energy: float
