+++ generatorcore/agri2018/energy_source/co2eFromEnergyUse.py
@@ -1,0 +1,23 @@
+# pyright: strict
+
+from dataclasses import dataclass
+
+from ..co2eEmissions import CO2eEmissions
+
+
+@dataclass(kw_only=True)
+class CO2eFromEnergyUse(CO2eEmissions):
+    # Used by s
+    CO2e_combustion_based: float
+    CO2e_production_based: float
+    CO2e_total: float
+    energy: float
+
+    @classmethod
+    def sum(cls, *co2es: "CO2eFromEnergyUse") -> "CO2eFromEnergyUse":  # type: ignore
+        return cls(
+            CO2e_combustion_based=sum(c.CO2e_combustion_based for c in co2es),
+            CO2e_production_based=sum(c.CO2e_production_based for c in co2es),
+            CO2e_total=sum(c.CO2e_total for c in co2es),
+            energy=sum(c.energy for c in co2es),
+        )
